@@ -3,6 +3,12 @@
   "private": true,
   "dependencies": {
     "@babel/core": "^7.23.0",
+    "@babel/plugin-transform-class-properties": "^7.23.3",
+    "@babel/plugin-transform-nullish-coalescing-operator": "^7.23.3",
+    "@babel/plugin-transform-numeric-separator": "^7.23.3",
+    "@babel/plugin-transform-object-rest-spread": "^7.23.3",
+    "@babel/plugin-transform-optional-chaining": "^7.23.3",
+    "@babel/plugin-transform-private-methods": "^7.23.3",
     "@babel/plugin-transform-runtime": "^7.22.15",
     "@babel/preset-env": "^7.22.20",
     "@babel/preset-react": "^7.22.15",
@@ -12,12 +18,16 @@
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@rails/webpacker": "5.4.4",
+    "@rollup/plugin-terser": "^0.4.4",
+    "autoprefixer": "^10.4.16",
     "aws-sdk": "^2.1467.0",
     "babel-loader": "8.0.4",
     "css-loader": "^5.2.7",
     "dotenv": "^16.3.1",
     "google-map-react": "^2.2.1",
     "moment": "^2.29.4",
+    "postcss": "^8.4.31",
+    "postcss-loader": "^7.3.3",
     "query-string": "^8.1.0",
     "react": "^18.2.0",
     "react-addons-shallow-compare": "^15.6.3",
@@ -26,19 +36,21 @@
     "react-dom": "^18.2.0",
     "react-geocode": "^0.2.3",
     "react-google-autocomplete": "^2.7.3",
-    "react-google-maps": "^9.4.5",
+    "react-google-maps": "^7.3.0",
     "react-hook-form": "^7.46.2",
     "react-places-autocomplete": "^7.3.0",
     "react-redux": "^8.1.2",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
+    "react-scripts": "^3.0.1",
     "react-with-direction": "^1.4.0",
     "react-with-styles": "^5.0.0",
     "redux": "^4.2.1",
     "redux-logger": "^3.0.6",
     "redux-thunk": "^2.4.2",
     "style-loader": "2.0.0",
-    "webpack": "^4.46.0",
+    "tailwindcss": "^3.3.5",
+    "webpack": "^5.89.0",
     "webpack-cli": "^3.3.12"
   },
   "description": "This README would normally document whatever steps are necessary to get the application up and running.",
@@ -51,11 +63,12 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "webpack -w --mode=development",
-    "postinstall": "webpack"
+    "postinstall": "webpack",
+    "build": "webpack postcss"
   },
   "engines": {
-    "node": "12.5.0",
-    "npm": "6.9.0"
+    "node": ">=12.5.0",
+    "npm": ">=6.9.0"
   },
   "keywords": [],
   "author": "",
@@ -63,6 +76,7 @@
   "devDependencies": {
     "@iconify/icons-simple-icons": "^1.2.70",
     "@iconify/react": "^4.1.1",
-    "webpack-dev-server": "^3"
+    "@webpack-cli/serve": "^2.0.5",
+    "webpack-dev-server": "^4.15.1"
   }
 }
